🦄 refactor(users): extract business logic into services, improve testability
- Extracted all business logic from views into dedicated service modules:
  users/services/{stats, notifications, export, groups, auth, profile, etc.}
- Simplified views.py by delegating logic to service functions
- Created modular test structure at users/tests/ for service-level testing
- Added type hints and optimized DB queries with select_related/prefetch_related
- Improved code readability, separation of concerns, and long-term maintainability

This refactoring aligns the app with Django best practices, enables easier testing,
and provides a clean foundation for future development.
