+++ hasta_la_vista_money/users/services/profile.py
@@ -1,0 +1,8 @@
+from hasta_la_vista_money.users.models import User
+
+
+def update_user_profile(form) -> User:
+    user = form.save(commit=False)
+    user.save()
+    form.save_m2m()
+    return user
