+++ hasta_la_vista_money/users/tests/test_registration.py
@@ -1,0 +1,29 @@
+from django.test import TestCase
+from django.contrib.auth import get_user_model
+from hasta_la_vista_money.users.services.registration import register_user
+from hasta_la_vista_money.users.forms import RegisterUserForm
+from django.utils import timezone
+
+User = get_user_model()
+
+
+class RegisterUserServiceTest(TestCase):
+    """Tests for register_user service function."""
+
+    def test_register_user(self):
+        data = {
+            'username': 'newuser',
+            'email': 'newuser@example.com',
+            'password1': 'StrongPassword123',
+            'password2': 'StrongPassword123',
+            'first_name': 'Test',
+            'last_name': 'User',
+        }
+        form = RegisterUserForm(data=data)
+        self.assertTrue(form.is_valid())
+        user = register_user(form)
+        self.assertTrue(user.is_superuser)
+        self.assertTrue(user.is_staff)
+        self.assertEqual(user.username, 'newuser')
+        self.assertEqual(user.email, 'newuser@example.com')
+        self.assertTrue(user.date_joined <= timezone.now())
