+++ setup.cfg
@@ -52,7 +52,7 @@ replace = __version__ = '{new_version}'
 universal = 1
 
 [mypy]
-python_version = 3.11
+python_version = 3.12
 
 files = hasta_la_vista_money/
